server/requirements.txt
fastapi
uvicorn
pyyaml
openai
graypy
Jinja2
alembic
sqlalchemy
asyncpg
databases
sqlalchemy
psycopg2-binary
python-multipart